+++ src/components/footer/index.js
@@ -24,3 +24,24 @@ const Footer = () => {
 };
 
 export default Footer;
+
+export const CommingSoonFooter = props => {
+  const date = new Date();
+
+  return (
+    <aside className={`footer__container ${props.allowScroll ? "relative_position" : ""}`}>
+      <footer className="main__footer">
+        <div className="socials">
+          <a href="https://twitter.com/thehexcord" className="icon__button"><span className="twitter"></span></a>
+          <a href="mailto:thehexcord@gmail.com" style={{ float: "right" }} className="icon__button"><span className="email"></span></a>
+        </div>
+        <div>
+          <NavLink to="/privacy-policy" className="boring__link">Privacy Policy</NavLink>
+        </div>
+        <div className="timestamp">
+          {date.toLocaleDateString()}
+        </div>
+      </footer>
+    </aside>
+  );
+};
