+++ src/styles/App.css
@@ -1,5 +1,5 @@
 .home__wrapper {
-  background: url(./assets/background.svg);
+  background: url(../assets/background.svg);
 }
 
 .container {
