+++ src/styles/typography.css
@@ -53,7 +53,7 @@
 
 
 .hyperlink {
-  font-family: sans-serif;
+  font-family: Lato;
   font-weight: normal;
   font-size: 13px;
   line-height: 18px;
